add packages, rename files
