+++ codegen.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+flutter packages pub run build_runner build --delete-conflicting-outputs
