+++ lib/classes/expense_detail.dart
