+++ lib/models/detail_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
-import 'package:budget_planner/classes/ExpenseDetail.dart';
+import 'package:budget_planner/classes/expense_detail.dart';
 
 class DetailCard extends StatefulWidget {
   late final List<ExpenseDetail> details;
