+++ pubspec.yaml
@@ -35,6 +35,11 @@ dependencies:
   shared_preferences: ^2.0.6
   async: ^2.6.1
   flutter_launcher_icons: ^0.9.0
+  # state management dependencies
+  mobx: ^2.0.1
+  mobx_codegen: ^2.0.1+3
+  beyond_helpers: ^2.0.0
+  build_runner: ^2.0.6
 
 dev_dependencies:
   flutter_test:
